Combine local_task_timing.py into task_timing.py
